+++ Source/VagabondsWork/Private/Visual/VagabondsWorldSettings.cpp
@@ -1,0 +1,6 @@
+#include "Visual/VagabondsWorldSettings.h"
+
+AVagabondsWorldSettings::AVagabondsWorldSettings()
+	: LevelVisualScheme(nullptr)
+{
+}