+++ Source/VagabondsWork/Ship.h
@@ -134,6 +134,9 @@ public:
     virtual void SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent) override;
     virtual void PossessedBy(AController* NewController) override;
     virtual void UnPossessed() override;
+
+    UPROPERTY(EditAnywhere, BlueprintReadWrite, Category="Ship|Custom")
+    bool Custom = false;
     
     virtual void NotifyHit(class UPrimitiveComponent* MyComp, class AActor* Other, class UPrimitiveComponent* OtherComp, bool bSelfMoved, FVector HitLocation, FVector HitNormal, FVector NormalImpulse, const FHitResult& Hit) override;